move interval subscription types to types file
