+++ src/controller/renderer/IntervalsController.ts
@@ -5,40 +5,10 @@ import { executeIntervaledOneShot } from '@/renderer/callbacks/intervaled';
 import { secondsUntilNextMinute } from '@/renderer/utils/timeUtils';
 import type { AnyData } from '@/types/misc';
 import type { ChainID } from '@/types/chains';
-import type { HelpItemKey } from '@/renderer/contexts/common/Help/types';
-
-/// Where `default` reads the tasks `enableOsNotifications` field.
-export type NotificationPolicy = 'default' | 'none' | 'one-shot';
-
-export interface IntervalSubscription {
-  // Unique id for the task.
-  action: string;
-  // Number of ticks between each one-shot execution.
-  intervalSetting: IntervalSetting;
-  // Used as a countdown.
-  tickCounter: number;
-  // Task category.
-  category: string;
-  // Task's associated chain.
-  chainId: ChainID;
-  // Shown in renderer.
-  label: string;
-  // Enabled or disabled.
-  status: 'enable' | 'disable';
-  // Flag to enable or silence OS notifications.
-  enableOsNotifications: boolean;
-  // Key to retrieve help information about the task.
-  helpKey: HelpItemKey;
-  // Associated referendum id for task.
-  referendumId?: number;
-  // Flag to determine if the subscription was just build (may not be needed)
-  justBuilt?: boolean;
-}
-
-export interface IntervalSetting {
-  label: string;
-  ticksToWait: number;
-}
+import type {
+  IntervalSetting,
+  IntervalSubscription,
+} from '@/types/subscriptions';
 
 export class IntervalsController {
   /// Active interval subscriptions keyed by chain ID.
