+++ src/renderer/contexts/main/IntervalSubscriptions/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 @rossbulat/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { IntervalSubscription } from '@/controller/renderer/IntervalsController';
+import type { IntervalSubscription } from '@/types/subscriptions';
 import type { ChainID } from '@/types/chains';
 
 export interface IntervalSubscriptionsContextInterface {
