+++ src/renderer/hooks/useMainMessagePorts.ts
@@ -34,7 +34,7 @@ import { useIntervalSubscriptions } from '../contexts/main/IntervalSubscriptions
 import type { AccountSource, LocalAddress } from '@/types/accounts';
 import type { ActiveReferendaInfo } from '@/types/openGov';
 import type { AnyData } from '@/types/misc';
-import type { IntervalSubscription } from '@/controller/renderer/IntervalsController';
+import type { IntervalSubscription } from '@/types/subscriptions';
 
 export const useMainMessagePorts = () => {
   /// Main renderer contexts.
