+++ src/types/subscriptions.ts
@@ -6,6 +6,39 @@ import type { AnyFunction, AnyData } from './misc';
 import type { FlattenedAccountData } from './accounts';
 import type { HelpItemKey } from '@/renderer/contexts/common/Help/types';
 
+/// Where `default` reads the tasks `enableOsNotifications` field.
+export type NotificationPolicy = 'default' | 'none' | 'one-shot';
+
+export interface IntervalSetting {
+  label: string;
+  ticksToWait: number;
+}
+
+export interface IntervalSubscription {
+  // Unique id for the task.
+  action: string;
+  // Number of ticks between each one-shot execution.
+  intervalSetting: IntervalSetting;
+  // Used as a countdown.
+  tickCounter: number;
+  // Task category.
+  category: string;
+  // Task's associated chain.
+  chainId: ChainID;
+  // Shown in renderer.
+  label: string;
+  // Enabled or disabled.
+  status: 'enable' | 'disable';
+  // Flag to enable or silence OS notifications.
+  enableOsNotifications: boolean;
+  // Key to retrieve help information about the task.
+  helpKey: HelpItemKey;
+  // Associated referendum id for task.
+  referendumId?: number;
+  // Flag to determine if the subscription was just build (may not be needed)
+  justBuilt?: boolean;
+}
+
 export type SubscriptionNextStatus = 'enable' | 'disable';
 
 export interface SubscriptionTask {
